+++ packages/unstyled/fab/src/Fab.tsx
@@ -34,6 +34,14 @@ function Fab<StyledFab>(StyledFab: React.ComponentType<StyledFab>) {
             disabled: isDisabled,
             focusVisible: isFocusVisibleProp || isFocusVisible,
           }}
+          dataSet={{
+            hover: isHoveredProp || isHovered ? 'true' : 'false',
+            focus: isFocusedProp || isFocused ? 'true' : 'false',
+            active: isPressedProp || isPressed ? 'true' : 'false',
+            disabled: isDisabled ? 'true' : 'false',
+            focusVisible:
+              isFocusVisibleProp || isFocusVisible ? 'true' : 'false',
+          }}
           disabled={isDisabled}
           {...(props as StyledFab)}
           onPressIn={composeEventHandlers(