+++ packages/unstyled/slider/src/Slider.tsx
@@ -133,6 +133,14 @@ function Slider<StyledSliderProps>(
               focusVisible: isFocusVisible || isFocusVisibleProp,
               active: isPressed || isPressedProp,
             }}
+            dataSet={{
+              hover: isHovered || isHoveredProp ? 'true' : 'false',
+              disabled: isDisabled || isDisabledProp ? 'true' : 'false',
+              focus: isFocused || isFocusedProp ? 'true' : 'false',
+              focusVisible:
+                isFocusVisible || isFocusVisibleProp ? 'true' : 'false',
+              active: isPressed || isPressedProp ? 'true' : 'false',
+            }}
             orientation={orientation ?? 'horizontal'}
             isReversed={isReversed ?? false}
           >