+++ packages/unstyled/menu/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @gluestack-ui/menu
 
+## 0.2.29
+
+### Patch Changes
+
+- fixes
+
+## 0.2.28
+
+### Patch Changes
+
+- Feature
+  Added Support of data-state-\* props
+
+## 0.2.27
+
+### Patch Changes
+
+- `Menu` fixed `closeOnSelect` onPress issue.
+
 ## 0.2.26
 
 ### Patch Changes